pyproject.toml
[project]
name = "bay-area-job-hunter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "python-jobspy>=1.1.82",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]
